Moved complex class into it's own class
