+++ src/com/jwetherell/algorithms/mathematics/FastFourierTransform.java
@@ -1,5 +1,7 @@
 package com.jwetherell.algorithms.mathematics;
 
+import com.jwetherell.algorithms.numbers.Complex;
+
 /**
  * A fast Fourier transform (FFT) algorithm computes the discrete Fourier transform (DFT) of a sequence, or its inverse. 
  * Fourier analysis converts a signal from its original domain (often time or space) to a representation in the frequency 
@@ -45,55 +47,6 @@ public class FastFourierTransform {
             Complex t = Complex.polar(1.0, -2 * Math.PI * k / size).multiply(odd[k]);
             coefficients[k] = even[k].add(t);
             coefficients[k + size / 2] = even[k].sub(t);
-        }
-    }
-
-    public static final class Complex {
-
-        public double real;
-        public double imaginary;
-
-        public Complex() {
-            this.real = 0.0;
-            this.imaginary = 0.0;
-        }
-
-        public Complex(double r, double i) {
-            this.real = r;
-            this.imaginary = i;
-        }
-
-        public Complex multiply(final Complex x) {
-            final Complex copy = new Complex(this.real, this.imaginary);
-            copy.real = this.real * x.real - this.imaginary * x.imaginary;
-            copy.imaginary = this.imaginary * x.real + this.real * x.imaginary;
-            return copy;
-        }
-
-        public Complex add(final Complex x) {
-            final Complex copy = new Complex(this.real, this.imaginary);
-            copy.real += x.real;
-            copy.imaginary += x.imaginary;
-            return copy;
-        }
-
-        public Complex sub(final Complex x) {
-            final Complex copy = new Complex(this.real, this.imaginary);
-            copy.real -= x.real;
-            copy.imaginary -= x.imaginary;
-            return copy;
-        }
-
-        public double abs() {
-            return Math.sqrt(this.real * this.real + this.imaginary * this.imaginary);
-        }
-
-        public String toString() {
-            return "(" + this.real + "," + this.imaginary + ")";
-        }
-
-        public static Complex polar(final double rho, final double theta) {
-            return (new Complex(rho * Math.cos(theta), rho * Math.sin(theta)));
         }
     }
 }
