+++ src/com/jwetherell/algorithms/mathematics/Multiplication.java
@@ -3,6 +3,8 @@ package com.jwetherell.algorithms.mathematics;
 import java.util.ArrayList;
 import java.util.Collections;
 
+import com.jwetherell.algorithms.numbers.Complex;
+
 public class Multiplication {
 
     public static final long multiplication(int a, int b) {
@@ -69,17 +71,17 @@ public class Multiplication {
         while (size < (a.length() + b.length())) {
             size *= 2;
         }
-        FastFourierTransform.Complex[] aCoefficients = new FastFourierTransform.Complex[size];
-        FastFourierTransform.Complex[] bCoefficients = new FastFourierTransform.Complex[size];
+        Complex[] aCoefficients = new Complex[size];
+        Complex[] bCoefficients = new Complex[size];
         for (int i = 0; i < size; i++) {
-            aCoefficients[i] = new FastFourierTransform.Complex();
-            bCoefficients[i] = new FastFourierTransform.Complex();
+            aCoefficients[i] = new Complex();
+            bCoefficients[i] = new Complex();
         }
         for (int i = 0; i < a.length(); i++) {
-            aCoefficients[i] = new FastFourierTransform.Complex((double) (Character.getNumericValue(a.charAt(a.length() - i - 1))), 0.0);
+            aCoefficients[i] = new Complex((double) (Character.getNumericValue(a.charAt(a.length() - i - 1))), 0.0);
         }
         for (int i = 0; i < b.length(); i++) {
-            bCoefficients[i] = new FastFourierTransform.Complex((double) (Character.getNumericValue(b.charAt(b.length() - i - 1))), 0.0);
+            bCoefficients[i] = new Complex((double) (Character.getNumericValue(b.charAt(b.length() - i - 1))), 0.0);
         }
 
         FastFourierTransform.cooleyTukeyFFT(aCoefficients);
@@ -89,7 +91,7 @@ public class Multiplication {
             aCoefficients[i] = aCoefficients[i].multiply(bCoefficients[i]);
         }
         for (int i = 0; i < size / 2; i++) {
-            FastFourierTransform.Complex temp = aCoefficients[i];
+            Complex temp = aCoefficients[i];
             aCoefficients[i] = aCoefficients[size - i - 1];
             aCoefficients[size - i - 1] = temp;
         }
